feat: integrate QueryEngine with Flask API and dynamic frontend
- Refactor `src/api/__init__.py` to use the `create_app` factory pattern and load `QueryEngine` as a singleton.
- Create 7 REST API endpoints in `src/api/routes.py` mapping to the 7 vector query modes, plus `/api/destinations` and `/api/health`.
- Update `src/web/templates/index.html` with a modern, dynamic UI using Tailwind CSS, featuring a sidebar form and a main results grid.
- Add `src/web/static/js/app.js` to handle form interactions, AJAX API calls, and render results dynamically as cards with similarity progress bars.
- Update `run.py` to correctly import and run the application factory.

+++ src/query_engine.py
@@ -52,7 +52,7 @@ class QueryEngine:
         Builds a secondary FAISS index with PCA dimensionality reduction (384 -> 64) for Q7.
         """
         self.pca = PCA(n_components=64)
-        reduced_embeddings = self.pca.fit_transform(self.embeddings)
+        reduced_embeddings = self.pca.fit_transform(self.embeddings).astype(np.float32)
 
         # Normalizing after PCA is necessary for Inner Product to represent Cosine Similarity
         reduced_embeddings = self.embedder.normalize(reduced_embeddings)
@@ -341,7 +341,7 @@ class QueryEngine:
         query_vec = self.embedder.embed_single(query).reshape(1, -1)
 
         # Transform using PCA
-        reduced_vec = self.pca.transform(query_vec)
+        reduced_vec = self.pca.transform(query_vec).astype(np.float32)
         reduced_vec = self.embedder.normalize(reduced_vec)
 
         distances, indices = self.indexer.search(self.pca_index, reduced_vec, k=k)
